Externalize compute duration
